refactor(top-bar): 컴포넌트명에 맞추어 파일명 변경
