+++ src/components/features/layout/top-bar/index.tsx
@@ -1,5 +1,5 @@
-import MainBar from './main';
-import PageBar from './page';
+import MainBar from './main-bar';
+import PageBar from './page-bar';
 
 // 이 곳에서 다 처리해야함
 // 1. Main 페이지에서의 Top - Bar
@@ -37,4 +37,5 @@ const TopBar = ({ type, ...props }: TopBarProps) => {
   }
   return null;
 };
+
 export default TopBar;
